refactor(@sounisi5011/encrypted-archive): stop using the filename `index.ts`
Node.js treats files named `index.*` specially.
In other words, we can omit the filename `index.*` when importing a directory path.
However, this is not possible on ESM.
For future migration to ESM, we will stop using the file name `index.ts`.
