+++ packages/encrypted-archive/src/core/header.ts
@@ -1,0 +1,2 @@
+export * from './header/create';
+export * from './header/parser';
