--- packages/encrypted-archive/src/core/header/index.ts
@@ -1,2 +1,0 @@
-export * from './create';
-export * from './parser';
