+++ packages/encrypted-archive/src/core/utils/reader.ts
@@ -1,6 +1,6 @@
-import { isOneArray, uint8arrayConcat } from '.';
 import type { BuiltinEncodeStringRecord, BuiltinInspectRecord } from '../types/builtin';
 import type { AsyncIterableIteratorReturn, AsyncIterableReturn } from '../types/utils';
+import { isOneArray, uint8arrayConcat } from '../utils';
 
 export interface BufferReaderInterface<T extends Uint8Array = Uint8Array> {
     read: (size: number, offset?: number) => Promise<T>;
